Consolidate the in-memory stream test harness
Six near-identical MemoryStream/MemoryStreamNetwork fakes (~960 LOC) had
drifted across slack/email/telegram/projects/scheduler suites, the GitHub
agent helpers, and a stream-repro subclass. They now all ride the one
canonical harness in streams/test-helpers.ts, which gains the three
capabilities the copies actually used:

- MemoryStreamNetwork: path-keyed streams for router suites observing
  cross-stream forwards (stream.at(path).append(...)), with an injectable
  clock feeding every member stream's createdAt (ack-freshness horizons)
  and a non-creating eventsAt (streams.size assertions stay honest).
- failAppendsOfType on MemoryStream: the scheduler's transient
  Stream-DO-outage injection.
- deliverNewEvents/ProcessorLike move in from agents/test-helpers.ts,
  which was itself a copy and is deleted.

Load-bearing divergences are kept as explicit pins, not silently
normalized: telegram and the GitHub agent suites pair processor clocks
(now: () => 60_000 / 999_999_999_999 / 10) with ~epoch createdAt stamps
for webhookAckIsFresh, so they subclass the network with super(() => 0)
and a comment saying why.

Zero test-semantics changes: 153 files / 1410 tests before and after.
