+++ apps/os/src/domains/agents/agent-processors.test.ts
@@ -1,5 +1,6 @@
 import { describe, expect, it, vi } from "vitest";
 import type { StreamEventInput } from "../streams/schemas.ts";
+import { MemoryStream, deliverNewEvents, type ProcessorLike } from "../streams/test-helpers.ts";
 import {
   AgentProcessor,
   buildAgentCompactionRequestBody,
@@ -16,7 +17,6 @@ import {
   DEFAULT_AGENT_MAX_AUTONOMOUS_TURNS,
   DEFAULT_AGENT_SYSTEM_PROMPT,
 } from "./agent-processor-contract.ts";
-import { MemoryStream, deliverNewEvents, type ProcessorLike } from "./test-helpers.ts";
 
 function agentRequestEvents(content: string, model = DEFAULT_AGENT_MODEL): StreamEventInput[] {
   return [
